Rename project contribution-related components.
Follow up to the discussion here: https://discord.com/channels/940363862723690546/940371255658418226/1024720356767965246
